# speeding up add function for large compounds