--- a/mbuild/compound.py
+++ b/mbuild/compound.py
@@ -870,13 +870,8 @@ class Compound(object):
             # compose children bond_graphs; make sure we actually have graphs to compose
             if len(temp_bond_graphs) != 0:
                 children_bond_graph = nx.compose_all(temp_bond_graphs)
-<<<<<<< HEAD
             
             if len(temp_bond_graphs) != 0 and not isinstance(self, Port) and children_bond_graph is not None:
-=======
-
-            if len(temp_bond_graphs) != 0 and not isinstance(self, Port):
->>>>>>> 3dfbfa514b377e5350be7b0e203f479dacc1ef74
                 # If anything is added at self level, it is no longer a particle
                 # search for self in self.root.bond_graph and remove self
                 if self.root.bond_graph.has_node(self):