--- a/mbuild/compound.py
+++ b/mbuild/compound.py
@@ -789,6 +789,7 @@ class Compound(object):
             if self.rigid_id:
                 if self.rigid_id > missing_rigid_id:
                     self.rigid_id -= 1
+
     # helper function to flattening a list that may be nested, in particular used for a list of labels
     def _flatten_list(self, c_list):
         if isinstance(c_list, list):
@@ -797,6 +798,7 @@ class Compound(object):
                     yield from self._flatten_list(c)
                 else:
                     yield c
+
     # helper function to flatten a list of Compounds that may be nested
     def _flatten_compound_list(self, c_list):
         if isinstance(c_list, Iterable):
@@ -871,7 +873,11 @@ class Compound(object):
             if len(temp_bond_graphs) != 0:
                 children_bond_graph = nx.compose_all(temp_bond_graphs)
 
-            if len(temp_bond_graphs) != 0 and not isinstance(self, Port) and children_bond_graph is not None:
+            if (
+                len(temp_bond_graphs) != 0
+                and not isinstance(self, Port)
+                and children_bond_graph is not None
+            ):
                 # If anything is added at self level, it is no longer a particle
                 # search for self in self.root.bond_graph and remove self
                 if self.root.bond_graph.has_node(self):