--- a/mbuild/compound.py
+++ b/mbuild/compound.py
@@ -511,7 +511,8 @@ class Compound(object):
         if self._contains_only_ports():
             return self._particle_mass(self)
         else:
-            particle_masses = [self._particle_mass(p) for p in self.particles()]
+            particle_masses = [self._particle_mass(
+                p) for p in self.particles()]
             if None in particle_masses:
                 warn(
                     f"Some particle of {self} does not have mass."
@@ -2506,7 +2507,8 @@ class Compound(object):
             if not isinstance(part, Compound):
                 raise MBuildError(f"{part} is not a Compound.")
             if id(part) != id(self) and id(part) not in successors_list:
-                raise MBuildError(f"{part} is not a member of Compound {self}.")
+                raise MBuildError(
+                    f"{part} is not a member of Compound {self}.")
 
             if check_if_particle:
                 if len(part.children) != 0:
@@ -2753,7 +2755,8 @@ class Compound(object):
 
             # Since the ignore_compounds can only be passed as a list
             # we can check the whole list at once before looping over it
-            self._check_openbabel_constraints(ignore_compounds, successors_list)
+            self._check_openbabel_constraints(
+                ignore_compounds, successors_list)
 
             for ignore in ignore_compounds:
                 p1 = ignore
@@ -3403,7 +3406,8 @@ class Compound(object):
             return list(self.particles())[selection]
         if isinstance(selection, str):
             if selection not in self.labels:
-                raise MBuildError(f"{self.name}['{selection}'] does not exist.")
+                raise MBuildError(
+                    f"{self.name}['{selection}'] does not exist.")
             return self.labels.get(selection)
 
     def __repr__(self):
@@ -3524,7 +3528,7 @@ Particle = Compound
 
 
 def _flatten_list(c_list):
-    """Flatten a list
+    """Flatten a list.
 
     Helper function to flatten a list that may be nested, e.g. [comp1, [comp2, comp3]].
     """