### PR Summary:
This refers to issue #1104 . This PR aims to improve the performance of the Compound.add function and loading routines that rely upon it. 

The basic gist, as outlined in the issue above, is that when constructing a compound using the add function, the performance can degrade as the compound grows in size, due to the repeated merging  (i.e., composing) of bond_graphs,  specifically, merging a small with a large bond graph over and over again.  This PR changes the underlying logic such that if a list of Compounds is passed to the add function, it will use the compose_all function to merge these bond_graphs together, before adding to the root Compound (and merging bond_graphs with the root compound).   The compose_all function effectively scales with the number of compounds being merged.  

This provides substantial speed improvements, as outlined in the issue. 

Other additions, Compound.add now accepts a list for the label argument if compounds are provided in a list. 

This is still a WIP, as tests need to be added for adding labels via a list, as well as adding in the updated load functions to stash compounds into lists (mdtraj conversion is basically complete and provides substantial speed up). 
  
### PR Checklist
------------
 - [x] Includes appropriate unit test(s)
 - [x] Appropriate docstring(s) are added/updated
 - [x] Code is (approximately) PEP8 compliant
 - [x] Issue(s) raised/addressed?