@chrisiacovella 2023-04-21
timings for loading mol2 files of different size:
|  N waters | new time (s) | old time (s) |
| -------- | ---------| -------- |
| 1000. | 0.347 | 4.92 |
| 5000 | 1.46 | 266.1  |
| 10000 | 3.02 |  too slow  |