@chrisiacovella 2023-04-22
Speed improvements for converting parmed are effectively the same as for mdtraj.  Note, converting from gmso will need modification still, but that calls a function in gmso, so it will need to be a separate gmso PR.