@chrisiacovella 2023-04-25
I added in a condense function to Compound.  This is similar to flatten, but it adds an intermediate level in the hierarchy based on connectivity.  This refers to issue #1108 . 

To reiterate the issue, take a compound that is like this:

`Compound, 28 particles, 18 bonds, 4 children
├── [C x 1], 1 particles, 0 bonds, 0 children
├── [Compound x 1], 3 particles, 2 bonds, 1 children
│   └── [tip3p x 1], 3 particles, 2 bonds, 3 children
│       ├── [H2 x 1], 1 particles, 1 bonds, 0 children
│       ├── [H3 x 1], 1 particles, 1 bonds, 0 children
│       └── [O1 x 1], 1 particles, 2 bonds, 0 children
└── [Compound x 2], 12 particles, 8 bonds, 4 children
    └── [Compound x 4], 3 particles, 2 bonds, 1 children
        └── [tip3p x 1], 3 particles, 2 bonds, 3 children
            ├── [H2 x 1], 1 particles, 1 bonds, 0 children
            ├── [H3 x 1], 1 particles, 1 bonds, 0 children
            └── [O1 x 1], 1 particles, 2 bonds, 0 children
`

And make it this:

`Compound, 28 particles, 18 bonds, 10 children
├── [C x 1], 1 particles, 0 bonds, 0 children
└── [tip3p x 9], 3 particles, 2 bonds, 3 children
    ├── [H2 x 1], 1 particles, 1 bonds, 0 children
    ├── [H3 x 1], 1 particles, 1 bonds, 0 children
    └── [O1 x 1], 1 particles, 2 bonds, 0 children
`

I still need to finish adding  in tests for this; that will come in the next push.