@chrisiacovella 2023-04-27
I addressed all the comments, including making one list_flatten helper function (doesn't add any real overhead).